PP-30 refactor: rename reducers
