+++ pointing-poker/src/redux/adapters/index.ts
@@ -1,18 +1,18 @@
 import { createEntityAdapter } from '@reduxjs/toolkit';
-import { Issue, IMessage, IUser } from '../../interfaces';
+import { Issue, Message, User } from '../../interfaces';
 
 export const issuesAdapter = createEntityAdapter<Issue>({
   selectId: issue => issue.id!,
 });
 
-export const messagesAdapter = createEntityAdapter<IMessage>({
+export const messagesAdapter = createEntityAdapter<Message>({
   selectId: message => message.id!,
 });
 
-export const usersAdapter = createEntityAdapter<IUser>({
+export const usersAdapter = createEntityAdapter<User>({
   selectId: user => user.id!,
 });
 
-export const newComerAdapter = createEntityAdapter<IUser>({
+export const newComerAdapter = createEntityAdapter<User>({
   selectId: newComer => newComer.id!,
 });
