+++ pointing-poker/src/redux/store/index.ts
@@ -6,11 +6,11 @@ import {
   messageReducer,
   issueReducer,
   currentUserReducer,
-  gameSettingsReducer,
+  settingsReducer,
   voteReducer,
   gameReducer,
   chatReducer,
-  gameRoundReducer,
+  roundReducer,
   newComerReducer,
   timerReducer,
 } from '../slices';
@@ -21,10 +21,10 @@ const appReducer = combineReducers({
   users: usersReducer,
   messages: messageReducer,
   issues: issueReducer,
-  gameSettings: gameSettingsReducer,
+  settings: settingsReducer,
   vote: voteReducer,
   chat: chatReducer,
-  round: gameRoundReducer,
+  round: roundReducer,
   newComers: newComerReducer,
   timer: timerReducer,
 });
